make draft of dbwindow
dbwindow is now spawned on click. placeholder labels and list widgets have been added.
in particular, creation is now added to a list in the mainwindow, and appropriately
despawned and reference freed from the list upon dbwindow closures. the mainwindow now also
automatically closes all dbwindows when it is itself closes.